requirements.txt
numpy==1.11.2
Pillow==3.4.2
scipy==0.18.1
wheel==0.26.0
